+++ Clients/src/presentation/components/Cards/RisksCard/style.ts
@@ -1,0 +1,27 @@
+export const projectRisksCard = {
+  border: `1px solid #eaecf0`,
+  borderRadius: 2,
+  backgroundColor: "#FFFFFF",
+  minWidth: "fit-content",
+  width: "fit-content",
+  display: "flex",
+  flexDirection: "row",
+  gap: 10,
+};
+
+export const projectRisksTileCard = {
+  paddingY: "15px",
+  paddingX: "30px",
+  textAlign: "center",
+  fontWeight: 600,
+  gap: 5,
+};
+
+export const projectRisksTileCardKey = {
+  fontSize: 13,
+};
+
+export const projectRisksTileCardvalue = {
+  fontSize: 28,
+  fontWeight: 800,
+};
