+++ Clients/src/presentation/components/Cards/TeamCard/style.ts
@@ -1,0 +1,21 @@
+export const infoCardStyle = {
+  border: `1px solid #eaecf0`,
+  borderRadius: 2,
+  backgroundColor: "#FFFFFF",
+  minWidth: 228,
+  width: "100%",
+  padding: "8px 36px 14px 14px",
+};
+
+export const infoCardTitleStyle = {
+  fontSize: 13,
+  color: "#8594AC",
+  pb: "2px",
+  textWrap: "wrap",
+};
+
+export const descCardbodyStyle = {
+  fontSize: 11,
+  color: "#2D3748",
+  textAlign: "justify",
+};
