+++ Clients/src/presentation/pages/ProjectView/V1.0ProjectView/Overview/style.ts
@@ -1,62 +1,6 @@
-export const infoCardStyle = {
-  border: `1px solid #eaecf0`,
-  borderRadius: 2,
-  backgroundColor: "#FFFFFF",
-  minWidth: 228,
-  width: "100%",
-  padding: "8px 36px 14px 14px",
-};
-
-export const infoCardTitleStyle = {
-  fontSize: 13,
-  color: "#8594AC",
-  pb: "2px",
-  textWrap: "wrap",
-};
-
-export const infoCardbodyStyle = {
-  fontSize: 16,
-  fontWeight: 600,
-  color: "#2D3748",
-};
-
-export const descCardbodyStyle = {
-  fontSize: 11,
-  color: "#2D3748",
-  textAlign: "justify",
-};
-
 export const rowStyle = {
   display: "flex",
   flexDirection: "row",
   gap: 10,
   mb: 10,
-};
-
-export const projectRisksCard = {
-  border: `1px solid #eaecf0`,
-  borderRadius: 2,
-  backgroundColor: "#FFFFFF",
-  minWidth: "fit-content",
-  width: "fit-content",
-  display: "flex",
-  flexDirection: "row",
-  gap: 10,
-};
-
-export const projectRisksTileCard = {
-  paddingY: "15px",
-  paddingX: "30px",
-  textAlign: "center",
-  fontWeight: 600,
-  gap: 5,
-};
-
-export const projectRisksTileCardKey = {
-  fontSize: 13,
-};
-
-export const projectRisksTileCardvalue = {
-  fontSize: 28,
-  fontWeight: 800,
 };
